Initial commit.
This first commit is a copy of the files from the Sourceforge project https://sourceforge.net/projects/ada-lapack/
There are no changes to the Ada source code. The reasons for the move are first, all of my other projects are on GitHub and second, I much prefer the GitHub environment to Sourceforge (the documentation is way better, the pages have *no* adverts and GitHub is by far the most popular open source hub).